refactor(数据大屏、仪表板): 优化图片组刷新
