+++ core/core-frontend/src/custom-component/picture-group/PictureGroupThreshold.vue
@@ -6,6 +6,7 @@ import Threshold from '@/views/chart/components/editor/editor-senior/components/
 import { CollapseSwitchItem } from '@/components/collapse-switch-item'
 import { useI18n } from '@/hooks/web/useI18n'
 import { useEmitt } from '@/hooks/web/useEmitt'
+import { ElMessage } from 'element-plus-secondary'
 const snapshotStore = snapshotStoreWithOut()
 const { t } = useI18n()
 
@@ -46,6 +47,18 @@ const onThresholdChange = val => {
   snapshotStore.recordSnapshotCache('calcData', view.value.id)
 }
 
+const onStyleChange = () => {
+  snapshotStore.recordSnapshotCache('onStyleChange')
+}
+
+const onRefreshChange = val => {
+  onStyleChange()
+  if (val === '' || parseFloat(val).toString() === 'NaN' || parseFloat(val) < 1) {
+    ElMessage.error(t('chart.only_input_number'))
+    return
+  }
+}
+
 const closeThreshold = () => {
   view.value.senior.threshold.enable = false
 }
@@ -73,6 +86,53 @@ onMounted(() => {
       :property-inner="['tableThreshold']"
       @onThresholdChange="onThresholdChange"
     />
+    <el-row v-if="view" class="refresh-area">
+      <el-form-item
+        style="width: 100%"
+        class="form-item no-margin-bottom"
+        :class="'form-item-' + themes"
+      >
+        <el-checkbox
+          v-model="view.refreshViewEnable"
+          :effect="themes"
+          size="small"
+          @change="onStyleChange()"
+        >
+          {{ t('visualization.refresh_frequency') }}
+        </el-checkbox>
+      </el-form-item>
+      <el-row style="width: 100%" v-if="view.refreshViewEnable">
+        <el-form-item
+          class="form-item no-margin-bottom select-append"
+          :class="'form-item-' + themes"
+        >
+          <el-input
+            v-model.number="view.refreshTime"
+            :effect="themes"
+            :class="[themes === 'dark' && 'dv-dark']"
+            size="small"
+            :min="1"
+            :max="3600"
+            :disabled="!view.refreshViewEnable"
+            @change="onRefreshChange"
+          >
+            <template #append>
+              <el-select
+                v-model="view.refreshUnit"
+                :effect="themes"
+                size="small"
+                placeholder="Select"
+                style="width: 80px"
+                @change="onStyleChange()"
+              >
+                <el-option :effect="themes" :label="t('visualization.minute')" :value="'minute'" />
+                <el-option :effect="themes" :label="t('visualization.second')" :value="'second'" />
+              </el-select>
+            </template>
+          </el-input>
+        </el-form-item>
+      </el-row>
+    </el-row>
   </collapse-switch-item>
 </template>
 
@@ -235,5 +295,9 @@ onMounted(() => {
   &:first-child {
     border-top: none !important;
   }
+}
+.refresh-area {
+  width: 100%;
+  padding: 0;
 }
 </style>
